pyproject.toml
[project]
name = "my_web"
version = "0.1.0"
description = "Flask skeleton: shows concepts, prevent repeating boilerplate code"
readme = "README.md"
authors = [
    { name = "Ondrej Profant", email = "ondrej.profant@gmail.com" }
]
requires-python = ">=3.13.0, <3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "ipython>=9.6.0",
    "pydantic-settings>=2.11.0",
]

[project.scripts]
web = "my_web.app:run"
shell = "my_web.app:shell"
help = "my_web.app:app_help"
load_fixtures = "my_web.app:load_fixtures"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.2",
]
